Move helper functions to tests.helper
