+++ tests/conftest.py
@@ -1,22 +1,7 @@
-import sys
-from contextlib import contextmanager
 from pathlib import Path
-from tempfile import NamedTemporaryFile, TemporaryDirectory
+from tempfile import TemporaryDirectory
 
 import pytest
-
-from magia import Elaborator, Module
-
-
-@contextmanager
-def elaborate_to_file(module: Module) -> str:
-    with NamedTemporaryFile(mode="w", suffix=".sv") as f:
-        f.write(Elaborator.to_string(module))
-        f.flush()
-        yield f.name
-
-
-sys.modules["pytest"].elaborate_to_file = elaborate_to_file
 
 
 @pytest.fixture()
